refactor: enable plugin and auth test
