+++ test/unit/modules/plugin/helper/plugin.js
@@ -2,11 +2,20 @@
 import path from 'path';
 import config from '../../../partials/config';
 
-const authProfileConf = {
-  ...config(),
+export const authProfileConf = config({
   auth: {
-    [`${path.join(__dirname, '../../../partials/plugin/authenticate')}`]: { }
+    [`${path.join(__dirname, '../../../partials/plugin/authenticate.success')}`]: { }
   }
-};
+});
 
-export default authProfileConf;
+export const authPluginFailureConf = config({
+  auth: {
+    [`${path.join(__dirname, '../../../partials/plugin/authenticate.fail')}`]: { }
+  }
+});
+
+export const authPluginPassThrougConf = config({
+  auth: {
+    [`${path.join(__dirname, '../../../partials/plugin/authenticate.passthroug')}`]: { }
+  }
+});
